Make a transitive closure that respects the ordering of a trace
Synchronous calls became problematic because abstracted the trace to only
contain scheduling events. The transitive closure of the abstracted
must-happen-before relation could lead to relating an event to itself, which
again would lead to two scheduling points in another cog to be related to each
other. This does not happen if we respect the ordering given by the trace.