Fix Source DPOR
The main issue was that a runtime can schedule events that are in the sleep
set, as it does not keep track of the sleep set. When this happens, disregard
the remaining of the trace, and add an event that is not in the sleep set in
the backtracking set. If no such event exists (i.e. the execution is sleep
blocked), then leave that branch unresolved.

Fixing the error also introduces a small change in the backtracking phase,
ensuring that we only reverse races where both events are within the part of
the execution that is not sleep blocked.